% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDomainParam-class.R
\docType{class}
\name{HRDomainParam-class}
\alias{HRDomainParam-class}
\title{Clase S4 HRFactorParam para los parámetros de la función IntervalUnitHitRate}
\description{
Definición de la clase S4 \code{HRDomainParam} que contiene los parámetros que utiliza la función
\code{\link{IntervalDomainHitRate}} para el cálculo del hit rate de un conjunto de intervalos
para un determinado dominio.
}
\section{Slots}{

\describe{
\item{\code{VarRoles}}{lista con las siguientes componentes:
\itemize{
\item \code{Units} Objeto de clase data.table con las unidades para las que se van a calcular
los intervalos.
\item \code{ObjVariable} Vector de tipo caracter de longitud 1 con el nombre de la variable
para la que se están construyendo los intervalos de validación.
\item \code{Domains} Vector con los nombres de las variables que actúan como factores.
\item \code{IntervalsLimits} Vector de longitud 2 con los nombres utilizados en el repositorio
para los extremos inferior y superior de los intervalos de validación.
\item \code{EditName} Vector de tipo caracter de longitud 1 con el nombre del edit para el
que se está construyendo los intervalos de validación.
}}
}}

\examples{
# Un prototipo vacío
new(Class = 'HRDomainParam')

}
