% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRFactor.R
\docType{methods}
\name{HRFactor}
\alias{HRFactor}
\alias{HRFactor,StQ,StQList,StQList,StQList,HRFactorParam-method}
\title{Cálculo del factor hit rate}
\usage{
HRFactor(object, EdData, RawData, IntervalData, Param)

\S4method{HRFactor}{StQ,StQList,StQList,StQList,HRFactorParam}(object, EdData,
  RawData, IntervalData, Param)
}
\arguments{
\item{object}{Objeto de clase \linkS4class{StQ} con los datos para los que se quiere calcular
el intervalo.}

\item{Param}{Objeto de clase \linkS4class{HRFactorParam} con los parámetros necesarios para
calcular el factor hit rate.}
}
\value{
\code{Vector} de tipo \code{numeric} con los factores hit rate de la variable.
}
\description{
\code{HRFactor} calcula el factor hit rate para obtener el radio de los intervalos de validación
con el método \code{\link{HRInterval}}.
}
\examples{
\dontrun{
HRFactor(object, Param)
}

}
