% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRUnitParam-class.R
\docType{class}
\name{HRUnitParam-class}
\alias{HRUnitParam-class}
\title{Clase S4 HRFactorParam para los parámetros de la función IntervalUnitHitRate}
\description{
Definición de la clase S4 \code{HRUnitParam} que contiene los parámetros que utiliza la función
\code{\link{IntervalUnitHitRate}} para el cálculo del hit rate de un conjunto de unidades.
}
\section{Slots}{

\describe{
\item{\code{VarRoles}}{lista con las siguientes componentes:

\item Units objeto de clase \code{data.table} con las unidades para las que se van a calcular
los intervalos.

\item ObjVariable \code{Vector} de tipo \code{character} de longitud 1 con el nombre de la variable
para la que se están construyendo los intervalos de validación.

\item IntervalsLimits \code{vector} de longitud 2 con los nombres utilizados en el repositorio
para los extremos inferior y superior de los intervalos de validación.}
}}

\examples{
# Un prototipo vacío
new(Class = 'HRUnitParam')

}
