% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalUnitHitRate.R
\docType{methods}
\name{IntervalUnitHitRate}
\alias{IntervalUnitHitRate}
\alias{IntervalUnitHitRate,StQ,StQList,StQList,StQList,HRUnitParam-method}
\title{Cálculo del hit rate para unidades}
\usage{
IntervalUnitHitRate(object, EdData, RawData, IntervalData, HRUnitParam)

\S4method{IntervalUnitHitRate}{StQ,StQList,StQList,StQList,HRUnitParam}(object,
  EdData, RawData, IntervalData, HRUnitParam)
}
\arguments{
\item{HRUnitParam}{Objeto de clase \linkS4class{HRUnitParam} con los parametros necesarios
para calcular el hit rate de un conjunto de unidades.}
}
\value{
\code{data.table} con el hit rate de los intervalos de entrada para cada unidad.
}
\description{
\code{IntervalUnitHitRate} calcula el hit rate de un conjunto de unidades.
}
\examples{
\dontrun{
IntervalUnitHitRate(HRUnitParam)
}

}
