% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateHRFactorParam-class.R
\docType{class}
\name{UpdateHRFactorParam-class}
\alias{UpdateHRFactorParam-class}
\title{Clase S4 para la actualización de los parámetros del Hit Rate por unidad}
\description{
Definición de la clase S4 llamada \code{UpdateHRFactorParam} para
la actualización de los parámetros utilizados en el cálculo del Hit Rate por
unidad en la fase longitudinal de la depuración selectiva.
}
\section{Slots}{

\describe{
\item{\code{EditName}}{Vector carácter con los nombres de los edits cuyos parámetros
queremos actualizar.}

\item{\code{Units}}{data.table con las unidades para las que se van a modificar los
parámetros..}

\item{\code{Values}}{lista con los valores de los parámetros que intervienen en el
cálculo del HitRate que se desean actualizar.
.}
}}

\examples{
new(Class = 'UpdateHRFactorParam')

}
