% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParamHRFactor.R
\docType{methods}
\name{UpdateParamHRFactor}
\alias{UpdateParamHRFactor}
\alias{UpdateParamHRFactor,data.table,UpdateHRFactorParam-method}
\alias{UpdateParamHRFactor,list,UpdateHRFactorParam-method}
\alias{UpdateParamHRFactor,list,list-method}
\title{Actualizacion de los parametros para el calculo del factor hit rate.}
\usage{
UpdateParamHRFactor(object, Param)

\S4method{UpdateParamHRFactor}{data.table,UpdateHRFactorParam}(object, Param)

\S4method{UpdateParamHRFactor}{list,UpdateHRFactorParam}(object, Param)

\S4method{UpdateParamHRFactor}{list,list}(object, Param)
}
\arguments{
\item{object}{Lista o \linkS4class{data.table} con los valores para cada edit
de los parametros que intervienen en el calculo del HitRate.}

\item{Param}{Objeto de clase \linkS4class{UpdateHRFactorParam} con las unidades
y valores de los parametros que se desean actualizar en el objeto de entrada
object.}
}
\value{
\code{Lista} con los parámetros para el calculo de los factores hit rate
del edit y unidades seleccionados actualizados.
}
\description{
\code{UpdateParamHRFactor} Actualiza la lista de parametros para el calculo del factor hit rate
con el metodo \code{\link{HRInterval}}.
}
\examples{
\dontrun{
UpdateHRFactorParam(object, Param)
}

}
