% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildUpdateHRFactorParam.R
\name{BuildUpdateHRFactorParam}
\alias{BuildUpdateHRFactorParam}
\title{Constructor de objetos de clase \linkS4class{UpdateHRFactorParam}}
\usage{
BuildUpdateHRFactorParam(EditName = character(0), Units = data.table(),
  Values = list())
}
\arguments{
\item{EditName}{\code{Vector} de tipo \code{character} de longitud 1 con el nombre
del Edit para el que se establece la condición.}

\item{Values}{\code{list} con los nuevos valores de los distintos parámetros del
HitRate que se quieren actualizar.}
}
\value{
Un objeto de clase \linkS4class{UpdateHRFactorParam} with components specified in the input
parameter Data. Depending on the specified components, the output object will be
of classes \code{GenConditionParam}, \code{FixedConditionParam} or \code{TSConditionParam}.
}
\description{
Este constructor devuelve un objeto de clase \linkS4class{UpdateHRFactorParam}.
a partir de un conjunto de parámetros.
}
\examples{
\dontrun{
UpdateHRFactorParam <- BuildUpdateHRFactorParam(EditName = 'LCN_W_1',
                                                )
UpdateHRFactorParam
}

}
