% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRFactorParam-class.R
\docType{class}
\name{HRFactorParam-class}
\alias{HRFactorParam-class}
\title{Clase S4 HRFactorParam para los parametros del método HRFactor}
\arguments{
\item{IntervalsLimits}{\code{vector} de longitud 2 con los nombres utilizados en el repositorio
para los extremos inferior y superior de los intervalos de validacion.}
}
\description{
Definicion de la clase S4 \code{HRFactorParam} que contiene los parametros que utiliza el metodo
\code{\link{HRFactor}} para el calculo del factor hit rate en la obtencion de los radios de los
intervalos de validacion con el metodo \code{\link{HRInterval}}.
}
\section{Slots}{

\describe{
\item{\code{Units}}{objeto de clase \code{data.table} con las unidades para las que se van a calcular
los intervalos.}

\item{\code{VarName}}{\code{Vector} de tipo \code{character} de longitud 1 con el nombre de la variable
para la que se estan construyendo los intervalos de validacion.}

\item{\code{Edit}}{\code{Vector} de tipo \code{character} de longitud 1 con el nombre del edit para el
que se esta construyendo los intervalos de validacion.}

\item{\code{DomainNames}}{\code{Vector} de tipo \code{character} con los nombres de la variables que
establecen las celdas (partición de la muestra).}

\item{\code{LastFactor}}{objeto de clase \code{data.table} con los valores del factor hit rate del edit
en el ultimo periodo, para cada unidad.}

\item{\code{MinFactor}}{objeto de clase \code{data.table} con los valores minimos del factor hit rate
de cada edit para cada unidad.}

\item{\code{MaxFactor}}{objeto de clase \code{data.table} con los valores maximos del factor hit rate
de cada edit para cada unidad.}

\item{\code{HRUnit}}{objeto de clase \code{data.table} con los valores del hit rate asociado al edit
especificado, para cada unidad.}

\item{\code{CHRUnit}}{objeto de clase \code{data.table} con los valores optimos de las unidades que no
han sido marcadas y efectivamente no tenian error, para el edit especificado y para cada unidad.}

\item{\code{HRDomain}}{objeto de clase \code{data.table} con los valores del hit rate asociado al edit
especificado, para cada celda (particion de la muestra).}

\item{\code{CHRDomain}}{objeto de clase \code{data.table} con los valores optimos de las unidades que no
han sido marcadas y efectivamente no tenian error, para el edit especificado y para cada celda
(partición de la muestra).}

\item{\code{HRlambda}}{objeto de clase \code{data.table} para la estabilizacion del factor asociado al
\emph{hit-rate}. Debe ser, para cada edit y cada unidad, un valor en el intervalo [0, 1] y esta
asociado al numero de unidades marcadas correctamente.}

\item{\code{CHRlambda}}{objeto de clase \code{data.table} para la estabilizacion del factor asociado al
\emph{hit-rate}. Debe ser, para cada edit y cada unidad, un valor en el intervalo [0, 1] y esta
asociado al numero de unidades no marcadas y sin error.}
}}

\examples{
# Un prototipo vacío
new(Class = 'HRFactorParam')

}
