% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRInterval.R
\docType{methods}
\name{HRInterval}
\alias{HRInterval}
\alias{HRInterval,StQList,StQList,StQList,HRIntervalParam-method}
\title{Calculo de intervalos de validacion}
\usage{
HRInterval(object, RawData, IntervalData, Param)

\S4method{HRInterval}{StQList,StQList,StQList,HRIntervalParam}(object, RawData,
  IntervalData, Param)
}
\arguments{
\item{object}{Objeto de clase \linkS4class{StQList} con el historico de datos depurados de las
unidades.}

\item{RawData}{objeto de clase \linkS4class{StQList} con el historico de datos sin depurar de las
unidades.}

\item{IntervalData}{objeto de clase \linkS4class{StQList} con el historico de los intervalos de
validacion de las unidades.}

\item{Param}{Objeto de clase \linkS4class{HRIntervalParam} con los parametros necesarios para
determinar los intervalos.}
}
\value{
\code{data.table} con los extremos inferior y superior de los intervalos de validacion
de cada unidad, asi como el factor hit rate asociado a la variable considerada para cada unidad.
}
\description{
\code{HRInterval} construye intervalos de validacion tomando como centro de los intervalos la
predicción de la variable considerada y como radio el producto de la desviación del error de
predicción por el factor hit rate asociado a la variable.
}
\examples{
\dontrun{
HRInterval(object, Param)
}

}
