% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRInterval.R
\docType{methods}
\name{HRInterval}
\alias{HRInterval}
\alias{HRInterval,StQList,StQList,StQList,HRIntervalParam-method}
\title{Cálculo de intervalos de validación}
\usage{
HRInterval(object, RawData, IntervalData, Param)

\S4method{HRInterval}{StQList,StQList,StQList,HRIntervalParam}(object, RawData,
  IntervalData, Param)
}
\arguments{
\item{object}{Objeto de clase \linkS4class{StQList} con los datos y unidades de las series de
tiempo con los que vamos a calcular los intervalos.}

\item{Param}{Objeto de clase \linkS4class{HRIntervalParam} con los parámetros necesarios para
determinar los intervalos.}
}
\value{
Objeto de clase \linkS4class{IntervalsDT} con los intervalos de validación para los datos
del objeto de entrada \code{object}.
}
\description{
\code{HRInterval} construye intervalos de validación determinando los valores de sus centros
y sus radios.
}
\examples{
\dontrun{
HRInterval(object, Param)
}

}
