% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRIntervalParam-class.R
\docType{class}
\name{HRIntervalParam-class}
\alias{HRIntervalParam-class}
\title{Clase S4 HRIntervalParam para los parametros del metodo HRInterval}
\arguments{
\item{IntervalsLimits}{\code{vector} de longitud 2 con los nombres utilizados en el repositorio
para los extremos inferior y superior de los intervalos de validacion.}
}
\description{
Definicion de la clase S4 \code{HRIntervalParam} que contiene los parametros que utiliza el
metodo \code{\link{HRInterval}}.
}
\section{Slots}{

\describe{
\item{\code{Units}}{objeto de clase \code{data.table} con las unidades para las que se van a calcular
los intervalos.}

\item{\code{VarName}}{\code{Vector} de tipo \code{character} de longitud 1 con el nombre de la variable
para la que se estan construyendo los intervalos de validacion.}

\item{\code{Edit}}{\code{Vector} de tipo \code{character} de longitud 1 con el nombre del edit para el
que se estan construyendo los intervalos de validacion.}

\item{\code{DomainNames}}{\code{Vector} de tipo \code{character} con los nombres de la variables que
establecen las celdas (particion de la muestra).}

\item{\code{LastFactor}}{objeto de clase \code{data.table} con los valores del factor hit rate del edit
en el ultimo periodo, para cada unidad.}

\item{\code{BestTSPred}}{Objeto de clase \linkS4class{data.table} con la mejor prediccion de entre un
conjunto de modelos para la variable considerada, para cada unidad.}

\item{\code{MinFactor}}{objeto de clase \code{data.table} con los valores minimos que ha de tener el
factor hit rate del edit para cada unidad.}

\item{\code{MaxFactor}}{objeto de clase \code{data.table} con los valores maximos que ha de tener el
factor hit rate del edit para cada unidad.}

\item{\code{HRUnit}}{objeto de clase \code{data.table} con los valores del hit rate asociado al edit
especificado, para cada unidad.}

\item{\code{CHRUnit}}{objeto de clase \code{data.table} con los valores optimos de las unidades que no
han sido marcadas y efectivamente no tenian error, para el edit especificado y para cada unidad}

\item{\code{HRDomain}}{objeto de clase \code{data.table} con los valores del hit rate de los intervalos
para cada celda (particion de la muestra).}

\item{\code{CHRDomain}}{objeto de clase \code{data.table} con los valores optimos de las unidades que no
han sido marcadas y efectivamente no tenian error para cada celda (particion de la muestra).}

\item{\code{HRlambda}}{objeto de clase \code{data.table} para la estabilizacion del factor asociado al
\emph{hit-rate}. Debe ser, para cada unidad, un valor en el intervalo [0, 1].}

\item{\code{CHRlambda}}{objeto de clase \code{data.table} para la estabilizacion del factor asociado al
\emph{hit-rate}. Debe ser, para cada unidad, un valor en el intervalo [0, 1] y esta asociado al
numero de unidades no marcadas y sin error.}

\item{\code{Lower}}{objeto de clase \code{data.table} que contiene, para cada unidad, valores fijos para 
los extremos inferiores de los intervalos o el valor NA en caso de que se quieran calcular.}

\item{\code{Upper}}{objeto de clase \code{data.table}que contiene, para cada unidad, valores fijos para 
los extremos superiores de los intervalos o el valor NA en caso de que se quieran calcular.}
}}

\examples{
# Un prototipo vacío
new(Class = 'HRIntervalParam')

}
