% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalUnitHitRate.R
\docType{methods}
\name{IntervalUnitHitRate}
\alias{IntervalUnitHitRate}
\alias{IntervalUnitHitRate,StQ,StQList,StQList,StQList,HRUnitParam-method}
\title{Calculo del hit rate para unidades}
\usage{
IntervalUnitHitRate(object, EdData, RawData, IntervalData, HRUnitParam)

\S4method{IntervalUnitHitRate}{StQ,StQList,StQList,StQList,HRUnitParam}(object,
  EdData, RawData, IntervalData, HRUnitParam)
}
\arguments{
\item{object}{objeto de clase \linkS4class{StQ} con las unidades para las
que se desea calcular el hit rate.}

\item{EdData}{objeto de clase \linkS4class{StQList} con el historico de
datos depurados de las unidades.}

\item{RawData}{objeto de clase \linkS4class{StQList} con el historico de
datos sin depurar de las unidades.}

\item{IntervalData}{objeto de clase \linkS4class{StQList} con el historico de
los intervalos de validacion de las unidades.}

\item{HRUnitParam}{Objeto de clase \linkS4class{HRUnitParam} con los parametros necesarios
para calcular el hit rate de un conjunto de unidades.}
}
\value{
\code{data.table} con el hit rate de los intervalos de entrada para cada unidad.
}
\description{
\code{IntervalUnitHitRate} calcula el hit rate de un conjunto de unidades.
}
\examples{
\dontrun{
IntervalUnitHitRate(HRUnitParam)
}

}
