% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateHRFactorParam-class.R
\docType{class}
\name{UpdateHRFactorParam-class}
\alias{UpdateHRFactorParam-class}
\title{Clase S4 para la actualizacion de los parametros del Hit Rate por unidad}
\description{
Definicion de la clase S4 llamada \code{UpdateHRFactorParam} para
la actualizacion de los parametros utilizados en el calculo del Hit Rate por
unidad en la fase longitudinal de la depuracion selectiva.
}
\section{Slots}{

\describe{
\item{\code{EditName}}{Vector caracter con los nombres de los edits cuyos parametros
queremos actualizar.}

\item{\code{Units}}{data.table con las unidades para las que se van a modificar los
parametros.}

\item{\code{Values}}{lista con los valores de los parametros que intervienen en el
cálculo del HitRate que se desean actualizar.
.}
}}

\examples{
new(Class = 'UpdateHRFactorParam')

}
