% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateHRFactorParam-class.R
\docType{class}
\name{UpdateHRFactorParam-class}
\alias{UpdateHRFactorParam-class}
\title{Clase S4 para la actualizacion de los parametros del Hit Rate por unidad}
\description{
Definicion de la clase S4 llamada \code{UpdateHRFactorParam} para la actualizacion
de los parametros utilizados en el calculo del Hit Rate por unidad en la fase longitudinal de la
depuracion selectiva.
}
\section{Slots}{

\describe{
\item{\code{EditName}}{\code{Vector} de tipo caracter y longitud 1 con el nombre del edit cuyos 
parametros se van a actualizar.}

\item{\code{Units}}{\code{data.table} con las unidades para las que se van a modificar los parametros.}

\item{\code{Values}}{lista con nombres que contiene los valores que actualizaran los parametros que 
intervienen en el calculo del HitRate por unidad.}
}}

\examples{
new(Class = 'UpdateHRFactorParam')

}
