% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParamHRFactor.R
\docType{methods}
\name{UpdateParamHRFactor}
\alias{UpdateParamHRFactor}
\alias{UpdateParamHRFactor,data.table,UpdateHRFactorParam-method}
\alias{UpdateParamHRFactor,list,UpdateHRFactorParam-method}
\alias{UpdateParamHRFactor,list,list-method}
\title{Actualizacion de los parametros para el calculo del factor hit rate.}
\usage{
UpdateParamHRFactor(object, Param)

\S4method{UpdateParamHRFactor}{data.table,UpdateHRFactorParam}(object, Param)

\S4method{UpdateParamHRFactor}{list,UpdateHRFactorParam}(object, Param)

\S4method{UpdateParamHRFactor}{list,list}(object, Param)
}
\arguments{
\item{object}{\linkS4class{data.table} con los valores de los parametros que intervienen en el
calculo del HitRate para un edit concreto, o lista de \linkS4class{data.table} con dichos
parametros para cada variable y cada edit.}

\item{Param}{Objeto de clase \linkS4class{UpdateHRFactorParam} con las unidades y valores de los
parametros que se desean actualizar en el objeto de entrada \code{object}, o lista de objetos de
este tipo cuando el objeto de entrada \code{object} es una lista.}
}
\value{
\linkS4class{data.table} o lista de \linkS4class{data.table}, segun sea el objeto de
entrada \code{object}, con los parametros para el calculo de los factores hit rate del edit y
unidades seleccionados actualizados.
}
\description{
\code{UpdateParamHRFactor} Actualiza la lista de parametros para el calculo del factor hit rate
con el metodo \code{\link{HRFactor}}.
}
\examples{
\dontrun{
UpdateHRFactorParam(object, Param)
}

}
